added jwt authentication

+++ src/main/java/com/example/FutoVoteBackend/repositories/StudentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.FutoVoteBackend.models.Student;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
 	Optional<Student> findByMatricNo(String matricNo);
 	Optional<Student> findByEmail(String email);
 }
